Refactor : UserMoney 클래스 PurchaseAmount 클래스로 변경
