--- src/main/java/lotto/domain/MoneyException.java
@@ -1,7 +1,0 @@
-package lotto.domain;
-
-public class MoneyException extends RuntimeException {
-    MoneyException(String message) {
-        super(message);
-    }
-}
