+++ src/main/java/lotto/domain/PurchaseAmountException.java
@@ -1,0 +1,7 @@
+package lotto.domain;
+
+public class PurchaseAmountException extends RuntimeException {
+    PurchaseAmountException(String message) {
+        super(message);
+    }
+}
